Modify bootstrap command to use https git URL
Our production server does not currently allow outgoing git connections. Due to updates in the salt bootstrap script, we are able to specify a git url to use with the `-g` flag.

We're using the 2014.1.0rc3 release of Salt to start implementing a change to how the mysql root password was handled